+++ build.sh
@@ -3,6 +3,8 @@ make run
 
 if test -d ./build;then
     echo "build dictory exist!"
+    rm -rf ./build
+    mkdir ./build
 else
     mkdir ./build
 fi
